fix: rm tokio from primitives
