+++ chains.example.json
@@ -1,0 +1,6 @@
+[
+    {
+        "rpc_url": "https://rpc.ankr.com/eth",
+        "vector_address": "0x0000000000000000000000000000000000000000"
+    }
+]
