+++ primitives/Cargo.toml
@@ -10,7 +10,6 @@ ed25519-consensus.workspace = true
 codec.workspace = true
 blake2.workspace = true
 itertools.workspace = true
-tokio.workspace = true
 alloy-primitives = "0.8"
 alloy-sol-types = { version = "0.8" }
