+++ primitives/src/lib.rs
@@ -120,23 +120,3 @@ mod tests {
         assert_eq!(extracted_hash, hash, "Hashes don't match")
     }
 }
-
-pub use timeout::Timeout;
-
-mod timeout {
-    use std::future::Future;
-    use std::time::Duration;
-    use tokio::time::{timeout, Timeout as TimeoutFuture};
-
-    pub trait Timeout: Sized {
-        fn timeout(self, duration: Duration) -> TimeoutFuture<Self>;
-    }
-
-    impl<T: Future> Timeout for T {
-        fn timeout(self, duration: Duration) -> TimeoutFuture<Self> {
-            timeout(duration, self)
-        }
-    }
-}
-
-mod maybe_signer {}
